feat(ai-sessions): replace the context panel with an assistant settings modal
