+++ frontend/src/lib/components/Section.svelte
@@ -102,7 +102,7 @@
 			transition:slide={animate || collapsable ? { duration: 200 } : { duration: 0 }}
 		>
 			{#if description}
-				<div class="text-xs text-primary mt-1 mb-2">{@html description}</div>
+				<div class="text-xs text-secondary mt-1 mb-2">{@html description}</div>
 			{/if}
 			<div class={twMerge('flex flex-col gap-6 grow min-h-0', headless ? '' : 'mt-4')}>
 				<div class={twMerge('grow min-h-0', clazz)}>
